Added all the files
This MAUI app is created for login attempts, it gives the user a combination of hints and five attempts to log in. If the user fails to log in within five attempts, then he/she is locked out. 
In MainPage.xaml: label, entry, and button are used. 
MainPage.xaml.cs: event handlers and logic of code(if-else condition) are used.
App.xaml.cs , AppShell.xaml & AppShell.xaml.cs files are also updated.